apply updated FileManagerModal
